build: move custom setuptools code to setup.py
